Migrate core stdpar allocation map data structure to common
